+++ lint.mjs
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+import { CONFIG_TYPE_STRING } from "config/src/CONFIG_TYPE.mjs";
 import { dirname } from "node:path";
 import { fileURLToPath } from "node:url";
 import { ShutdownHandler } from "shutdown-handler/src/ShutdownHandler.mjs";
@@ -6,11 +7,6 @@ import { ShutdownHandler } from "shutdown-handler/src/ShutdownHandler.mjs";
 const shutdown_handler = await ShutdownHandler.new();
 
 try {
-    const path = process.argv[2] ?? null;
-    if (path === null) {
-        throw new Error("Please pass a path!");
-    }
-
     const eslint = new (await import("eslint")).ESLint({
         cwd: dirname(fileURLToPath(import.meta.url)),
         errorOnUnmatchedPattern: false,
@@ -21,11 +17,18 @@ try {
             ".mjs"
         ],
         globInputPaths: false,
-        overrideConfig: (await import("../.eslintrc.json", { with: { type: "json" } })).default,
+        overrideConfig: (await import("./.eslintrc.json", { with: { type: "json" } })).default,
         useEslintrc: false
     });
 
-    const result = (await eslint.loadFormatter()).format(await eslint.lintFiles(path));
+    const result = (await eslint.loadFormatter()).format(await eslint.lintFiles(await (await (await import("config/src/Config.mjs")).Config.new(
+        await (await import("config/src/getValueProviders.mjs")).getValueProviders(
+            true
+        )
+    )).getConfig(
+        "path",
+        CONFIG_TYPE_STRING
+    )));
 
     console.log(result);