+++ pre-publish.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env sh
+
+set -e
+
+root_folder="`dirname "$0"`"
+
+#"$root_folder/lint.sh"
+
+publish-utils-update-release-version "$root_folder"
+
+"$root_folder/build.sh"