+++ CHANGELOG.md
@@ -6,6 +6,14 @@ Changes:
 
 \-
 
+## v2024-05-28-1
+
+Changes:
+
+- Remove `bin` folder
+- Seperate docker
+- `application-id` build config
+
 ## v2024-05-08-1
 
 Changes: